refactor(structure): reorganize backend into feature-based folders for better modularity and maintainability
