+++ backend/features/assignments/assignments.routes.js
@@ -1,8 +1,8 @@
 const express = require('express');
 const { body, validationResult } = require('express-validator');
-const Assignment = require('../models/Assignment');
-const Course = require('../models/Course');
-const { auth, authorize, checkApproval } = require('../middleware/auth');
+const Assignment = require('./Assignment');
+const Course = require('../courses/Course');
+const { auth, authorize, checkApproval } = require('../../middleware/auth');
 
 const router = express.Router();
 
@@ -15,7 +15,7 @@ router.get('/', auth, async (req, res) => {
     
     if (req.user.role === 'student') {
       // Get assignments from enrolled courses
-      const Enrollment = require('../models/Enrollment'); // Fixed: Move require inside function
+      const Enrollment = require('../enrollments/Enrollment'); // Fixed: Move require inside function
       const enrollments = await Enrollment.find({ 
         student: req.user._id,
         status: 'enrolled'
@@ -119,8 +119,8 @@ router.post('/', [
 
     // If assignment is published, notify enrolled students
     if (isPublished) {
-      const Enrollment = require('../models/Enrollment');
-      const Notification = require('../models/Notification');
+      const Enrollment = require('../enrollments/Enrollment');
+      const Notification = require('../notifications/Notification');
       
       const enrolledStudents = await Enrollment.find({
         course: courseId,
