+++ backend/features/notifications/notifications.routes.js
@@ -1,6 +1,6 @@
 const express = require('express');
-const Notification = require('../models/Notification');
-const { auth } = require('../middleware/auth');
+const Notification = require('./Notification');
+const { auth } = require('../../middleware/auth');
 
 const router = express.Router();
